backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt==4.0.1
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
python-dotenv
alembic
pydantic[email]
cors
